requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.34
asyncpg==0.29.0
pydantic==2.9.0
pydantic-settings==2.4.0
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0

